+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## v3.1.6f - 2022-02-XX
 New features:
+ - Make `geographies` field required [#777](https://github.com/okfn/ckanext-unhcr/pull/777)
  - Update label for _Pop of concern_ to _Pop. of Concern Location_ [#780](https://github.com/okfn/ckanext-unhcr/pull/780)
  - Minimum search for `geographies` to 2 characters [#779](https://github.com/okfn/ckanext-unhcr/pull/779)
  - Cache KoBo requests [#787](https://github.com/okfn/ckanext-unhcr/pull/787)